requirements.txt
pyautogui
pillow
opencv-python
numpy
pytesseract
